keepalive updates
git-svn-id: svn://svn.caucho.com/resin/trunk@1612 9c94448d-38f1-0310-a231-d98308ff1ebf

+++ modules/resin/src/com/caucho/db/store/Store.java
@@ -111,7 +111,7 @@ public class Store {
   public final static long BLOCK_OFFSET_MASK = BLOCK_SIZE - 1;
 
   public final static int ALLOC_PER_BLOCK = BLOCK_SIZE / 2;
-  public final static long ALLOC_SEGMENT = BLOCK_SIZE * ALLOC_PER_BLOCK;
+  public final static long ALLOC_SEGMENT = (long) BLOCK_SIZE * ALLOC_PER_BLOCK;
   
   public final static int ALLOC_FREE     = 0x00;
   public final static int ALLOC_ROW      = 0x01;
